outsource utils into header
